// src/types/structure/postgres-data-type.ts
// Data type info as returned from PG
export type PostgresDataType =
  | "bigint"
  | "bigserial"
  | "bit"
  | "bit varying"
  | "boolean"
  | "box"
  | "bytea"
  | "character"
  | "character varying"
  | "cidr"
  | "circle"
  | "date"
  | "double precision"
  | "inet"
  | "integer"
  | "interval"
  | "json"
  | "jsonb"
  | "line"
  | "lseg"
  | "macaddr"
  | "macaddr8"
  | "money"
  | "numeric"
  | "path"
  | "pg_lsn"
  | "pg_snapshot"
  | "point"
  | "polygon"
  | "real"
  | "smallint"
  | "smallserial"
  | "serial"
  | "text"
  | "time"
  | "time with time zone"
  | "timestamp"
  | "timestamp with time zone"
  | "tsquery"
  | "tsvector"
  | "txid_snapshot"
  | "uuid"
  | "xml";